refactor(db): extract resolver projection glue into sibling modules
Move the canonical-entity-resolution-specific pieces of graph.rs and
queries.rs into two new sibling modules, mirroring the existing
graph_inventory.rs / graph_findings.rs pattern instead of leaving both
files on the rust-module-size.allow exemption list.

src/db/graph_resolver_projection.rs (+ sidecar tests):
- extract_agent_docker_row, agent_docker_observations_from_log_row,
  project_resolver_decisions, trust_to_graph
- graph_walk_service_topic and its GRAPH_SERVICE_TOPIC_* constants
- cleanup_legacy_service_topology

src/db/queries_service_instances.rs (+ sidecar tests):
- search_logs_for_service_instances, service_instances_of_logical_services,
  search_logs_by_hostnames, run_union_all_log_arms, log_window_filter_tail,
  escape_like
- resolve_topic_entities, glob_prefix_pattern (indexed-tier topic
  resolution, syslog-mcp-csukc)

topic_correlate_inputs stays in queries.rs as the integration point tying
these helpers together with the general-purpose graph walk / log fan-out.
EntityMemo, ensure_entity_memoized, LogGraphRow, and the rest of the
extract_*_row dispatch family stay in graph.rs since wave 3
(syslog-mcp-sfm5o) made that memo shared across every extractor, not just
the resolver path — extracting it would drag along the whole dispatch tree.

Pure structural move: shared machinery gained pub(crate)/pub(super)
visibility so the new modules can call into it, call sites were updated
across queries.rs, graph_inventory.rs, and db.rs's re-export surface, and
directly-affected unit tests moved to the new modules' sidecar test files.
No behavior change. graph.rs (3007 -> 2759 non-comment lines) and
queries.rs (3230 -> 2957) both still exceed the 500-line gate and keep
their rust-module-size.allow entries; the two new modules (261 and 298
non-comment lines) pass cleanly and were not added to the allowlist.
