+++ src/db/pool_tests.rs
@@ -2083,102 +2083,6 @@ fn graph_schema_accepts_entity_resolution_vocabulary() {
 }
 
 #[test]
-fn stale_service_topology_cleanup_removes_old_canonical_rows() {
-    let dir = tempfile::tempdir().unwrap();
-    let pool = init_pool(&StorageConfig::for_test(
-        dir.path().join("stale-service-cleanup.db"),
-    ))
-    .unwrap();
-    let mut conn = pool.get().unwrap();
-    conn.execute(
-        "INSERT INTO graph_entities
-            (entity_type, canonical_key, display_label, source_kind, source_id, trust_level)
-         VALUES
-            ('service', 'tootie:plex', 'plex', 'log', 'fixture', 'inferred'),
-            ('service', 'tootie:plex:plex', 'tootie/plex/plex', 'log', 'fixture', 'inferred'),
-            ('app', 'plex/plex/plex', 'plex/plex/plex', 'log', 'fixture', 'claimed')",
-        [],
-    )
-    .unwrap();
-    // Seed more legacy rows than one cleanup chunk (2000) so the chunked
-    // delete loop must iterate, plus dependent alias/relationship/evidence
-    // rows to exercise every phase, and one unrelated host that must survive.
-    {
-        let tx = conn.transaction().unwrap();
-        for i in 0..2_500 {
-            tx.execute(
-                "INSERT INTO graph_entities
-                    (entity_type, canonical_key, display_label, source_kind, source_id, trust_level)
-                 VALUES ('service', ?1, 'svc', 'log', 'fixture', 'inferred')",
-                [format!("host{i}:svc{i}")],
-            )
-            .unwrap();
-        }
-        tx.execute(
-            "INSERT INTO graph_entities
-                (entity_type, canonical_key, display_label, source_kind, source_id, trust_level)
-             VALUES ('host', 'tootie', 'tootie', 'log', 'fixture', 'verified')",
-            [],
-        )
-        .unwrap();
-        tx.execute_batch(
-            "INSERT INTO graph_entity_aliases
-                (entity_id, alias_type, alias_key, alias_value, source_kind, trust_level)
-             SELECT id, 'app_name', canonical_key, canonical_key, 'log', 'inferred'
-               FROM graph_entities WHERE entity_type = 'service';
-             INSERT INTO graph_relationships
-                (relationship_key, src_entity_id, dst_entity_id, relationship_type,
-                 reason_code, trust_level, confidence)
-             SELECT s.id || ':runs_on:' || h.id, s.id, h.id, 'runs_on',
-                    'docker_service_label', 'inferred', 0.5
-               FROM graph_entities s, graph_entities h
-              WHERE s.entity_type = 'service' AND h.canonical_key = 'tootie';
-             INSERT INTO graph_relationship_evidence
-                (relationship_id, evidence_key, source_kind, source_id,
-                 observed_at, reason_code, trust_level)
-             SELECT id, 'ev:' || id, 'log', 'fixture',
-                    '2026-01-01T00:00:00Z', 'docker_service_label', 'inferred'
-               FROM graph_relationships;",
-        )
-        .unwrap();
-        tx.commit().unwrap();
-    }
-    crate::db::graph::cleanup_legacy_service_topology(&mut conn).unwrap();
-    let count: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM graph_entities
-              WHERE entity_type = 'service'
-                 OR (entity_type = 'app' AND canonical_key = 'plex/plex/plex')",
-            [],
-            |row| row.get(0),
-        )
-        .unwrap();
-    assert_eq!(count, 0);
-    // Dependent rows are fully gone across every chunked phase.
-    for table in [
-        "graph_entity_aliases",
-        "graph_relationships",
-        "graph_relationship_evidence",
-    ] {
-        let n: i64 = conn
-            .query_row(&format!("SELECT COUNT(*) FROM {table}"), [], |row| {
-                row.get(0)
-            })
-            .unwrap();
-        assert_eq!(n, 0, "{table} must be emptied by cleanup");
-    }
-    // The unrelated host entity survives the cleanup.
-    let hosts: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM graph_entities WHERE canonical_key = 'tootie'",
-            [],
-            |row| row.get(0),
-        )
-        .unwrap();
-    assert_eq!(hosts, 1);
-}
-
-#[test]
 fn migration_41_cleans_legacy_service_rows_from_populated_db() {
     // Simulate a populated pre-41 DB: run all migrations, then re-insert the
     // old-shaped rows a v40 DB could contain and re-run the 41 cleanup SQL by
