+++ src/db/queries_graph_tests.rs
@@ -367,199 +367,6 @@ fn topic_correlate_app_seed_does_not_fan_out_to_whole_host() {
 }
 
 #[test]
-fn search_logs_for_service_instances_uses_service_predicates_not_host_fanout() {
-    let (_dir, pool) = test_pool("service-instance-predicates.db");
-    let mut plex = syslog_row("2026-01-01T00:01:00Z", "tootie", "plex/plex/plex");
-    plex.metadata_json = Some(
-        r#"{"source_kind":"agent-docker","agent_docker":{"host":"tootie","container_id":"abc","container_name":"plex","compose_service":"plex","stream":"stdout"}}"#
-            .to_string(),
-    );
-    let mut exact = syslog_row("2026-01-01T00:02:00Z", "tootie", "plex");
-    exact.metadata_json = None;
-    insert_logs_batch(
-        &pool,
-        &[
-            syslog_row("2026-01-01T00:00:00Z", "tootie", "kernel"),
-            plex,
-            exact,
-            syslog_row("2026-01-01T00:03:00Z", "shart", "plex"),
-        ],
-    )
-    .unwrap();
-
-    let rows = search_logs_for_service_instances(
-        &pool,
-        &["tootie/plex".to_string()],
-        None,
-        None,
-        None,
-        50,
-    )
-    .unwrap();
-    // Matches: exact app label, prefixed nested label, structured compose
-    // service — all scoped to the instance's host. The kernel row on the
-    // same host and the other host's plex row stay out.
-    assert_eq!(rows.len(), 2);
-    assert!(rows.iter().all(|row| row.entry.hostname == "tootie"));
-    assert!(
-        rows.iter()
-            .all(|row| row.entry.app_name.as_deref() != Some("kernel"))
-    );
-    assert!(
-        rows.iter()
-            .all(|row| row.inclusion_reason == "service_instance")
-    );
-    assert!(
-        rows.iter()
-            .all(|row| row.resolver_status == ResolverStatus::Resolved)
-    );
-    assert!(rows.iter().all(|row| row.fallback_kind.is_none()));
-}
-
-#[test]
-fn search_logs_for_service_instances_escapes_like_wildcards() {
-    let (_dir, pool) = test_pool("service-instance-like-escape.db");
-    insert_logs_batch(
-        &pool,
-        &[
-            // `_` is a LIKE single-char wildcard; an unescaped pattern
-            // `my_app/%` would match this row.
-            syslog_row("2026-01-01T00:00:00Z", "tootie", "myxapp/x"),
-            syslog_row("2026-01-01T00:01:00Z", "tootie", "my_app/x"),
-        ],
-    )
-    .unwrap();
-    let rows = search_logs_for_service_instances(
-        &pool,
-        &["tootie/my_app".to_string()],
-        None,
-        None,
-        None,
-        50,
-    )
-    .unwrap();
-    assert_eq!(rows.len(), 1);
-    assert_eq!(rows[0].entry.app_name.as_deref(), Some("my_app/x"));
-}
-
-#[test]
-fn service_instance_fanout_arms_use_index_search_without_temp_btree() {
-    let (_dir, pool) = test_pool("service-instance-plan.db");
-    let conn = pool.get().unwrap();
-    // Two-arm UNION ALL replica of the search_logs_for_service_instances
-    // shape: every arm must be an index search (no `SCAN logs`) and no arm
-    // may sort through a temp b-tree — LIMIT pushdown streams each arm off
-    // idx_logs_host_time in timestamp order.
-    let arm = "SELECT * FROM (SELECT l.id
-          FROM logs l
-         WHERE l.hostname = ? AND (l.app_name = ? OR l.app_name LIKE ? ESCAPE '\\' \
-        OR json_extract(l.metadata_json, '$.agent_docker.compose_service') = ?)
-         ORDER BY l.timestamp DESC, l.id DESC
-         LIMIT ?)";
-    let sql = format!("EXPLAIN QUERY PLAN {arm} UNION ALL {arm}");
-    let plan: Vec<String> = conn
-        .prepare(&sql)
-        .unwrap()
-        .query_map(
-            rusqlite::params![
-                "tootie", "plex", "plex/%", "plex", 100, "shart", "plex", "plex/%", "plex", 100
-            ],
-            |row| row.get::<_, String>(3),
-        )
-        .unwrap()
-        .collect::<rusqlite::Result<Vec<_>>>()
-        .unwrap();
-    assert!(
-        plan.iter()
-            .any(|p| p.contains("USING INDEX idx_logs_host_time")),
-        "arms must search idx_logs_host_time: {plan:?}"
-    );
-    assert!(
-        !plan.iter().any(|p| p.starts_with("SCAN logs")),
-        "no arm may full-scan logs: {plan:?}"
-    );
-    assert!(
-        !plan.iter().any(|p| p.contains("TEMP B-TREE")),
-        "no temp b-tree sort allowed: {plan:?}"
-    );
-}
-
-#[test]
-fn search_logs_for_service_instances_rejects_legacy_keys() {
-    let (_dir, pool) = test_pool("service-instance-legacy-keys.db");
-    insert_logs_batch(
-        &pool,
-        &[syslog_row("2026-01-01T00:00:00Z", "tootie", "plex")],
-    )
-    .unwrap();
-    // Legacy shapes never split into (host, service) and yield no predicate.
-    let rows = search_logs_for_service_instances(
-        &pool,
-        &["tootie:plex".to_string(), "plex/plex/plex".to_string()],
-        None,
-        None,
-        None,
-        50,
-    )
-    .unwrap();
-    assert!(rows.is_empty());
-}
-
-#[test]
-fn graph_walk_service_topic_traverses_proof_edges_and_caps_results() {
-    let (_dir, pool) = test_pool("service-topic-walk.db");
-    let conn = pool.get().unwrap();
-    let logical = insert_entity(&conn, "logical_service", "plex");
-    let instance = insert_entity(&conn, "service_instance", "tootie/plex");
-    let host = insert_entity(&conn, "host", "tootie");
-    let app = insert_entity(&conn, "app", "kernel");
-    insert_rel(&conn, instance, logical, "instance_of");
-    insert_rel(&conn, instance, host, "runs_on");
-    // Broad log-identity edge from the host: must NOT be traversed.
-    insert_rel(&conn, app, host, "emitted_by");
-
-    let (entities, truncated) =
-        graph::graph_walk_service_topic(&conn, &["plex".to_string()], 3).unwrap();
-    let walked = keys(&entities);
-    assert!(walked.contains(&"plex".to_string()));
-    assert!(walked.contains(&"tootie/plex".to_string()));
-    assert!(walked.contains(&"tootie".to_string()));
-    assert!(
-        !walked.contains(&"kernel".to_string()),
-        "service-topic walk must not traverse broad log-identity edges: {walked:?}"
-    );
-    // Exact reach: seed + instance + host, and nothing else. (A `<= CAP`
-    // assertion would be vacuous on a 4-node fixture.)
-    assert_eq!(entities.len(), 3, "walk must reach exactly {walked:?}");
-    assert!(!truncated, "small fixture must not report truncation");
-}
-
-#[test]
-fn graph_walk_service_topic_reports_truncated_when_cap_hit() {
-    let (_dir, pool) = test_pool("service-topic-walk-cap.db");
-    let conn = pool.get().unwrap();
-    let seed = insert_entity(&conn, "logical_service", "plex");
-    // One more neighbor than GRAPH_SERVICE_TOPIC_ENTITY_CAP so the walk
-    // (seed + neighbors) exceeds the cap and must report truncation.
-    for i in 0..(graph::GRAPH_SERVICE_TOPIC_ENTITY_CAP + 1) {
-        let neighbor = insert_entity(&conn, "host", &format!("host-{i}"));
-        insert_rel(&conn, seed, neighbor, "runs_on");
-    }
-
-    let (entities, truncated) =
-        graph::graph_walk_service_topic(&conn, &["plex".to_string()], 1).unwrap();
-    assert_eq!(
-        entities.len(),
-        graph::GRAPH_SERVICE_TOPIC_ENTITY_CAP,
-        "result must be capped at GRAPH_SERVICE_TOPIC_ENTITY_CAP"
-    );
-    assert!(
-        truncated,
-        "walk reaching more than the cap must report truncated=true"
-    );
-}
-
-#[test]
 fn ambiguous_prefix_candidates_surface_without_log_fanout() {
     let (_dir, pool) = test_pool("topic-ambiguous-prefix.db");
     insert_logs_batch(
@@ -590,179 +397,5 @@ fn ambiguous_prefix_candidates_surface_without_log_fanout() {
     assert!(
         inputs.logs.is_empty(),
         "ambiguous candidates must contribute zero log fan-out"
-    );
-}
-
-#[test]
-fn service_instance_fanout_truncates_globally_newest_first_across_arms() {
-    let (_dir, pool) = test_pool("service-instance-union-truncation.db");
-    // Two instances on different hosts, limit + 2 rows each, sharing the
-    // same timestamp set so the merge exercises the tie-break.
-    let mut rows = Vec::new();
-    for i in 0..8 {
-        let ts = format!("2026-01-01T00:00:{i:02}Z");
-        rows.push(syslog_row(&ts, "tootie", "plex"));
-        rows.push(syslog_row(&ts, "shart", "plex"));
-    }
-    insert_logs_batch(&pool, &rows).unwrap();
-
-    let limit = 6;
-    let out = search_logs_for_service_instances(
-        &pool,
-        &["tootie/plex".to_string(), "shart/plex".to_string()],
-        None,
-        None,
-        None,
-        limit,
-    )
-    .unwrap();
-    // UNION ALL arms each fetch up to `limit`; the Rust merge must truncate
-    // to exactly `limit` rows globally.
-    assert_eq!(out.len(), limit);
-    // Global newest-first with the (timestamp DESC, id DESC) tie-break.
-    for pair in out.windows(2) {
-        let (a, b) = (&pair[0].entry, &pair[1].entry);
-        assert!(
-            a.timestamp > b.timestamp || (a.timestamp == b.timestamp && a.id > b.id),
-            "rows must be (timestamp DESC, id DESC): {}#{} then {}#{}",
-            a.timestamp,
-            a.id,
-            b.timestamp,
-            b.id
-        );
-    }
-    // The globally newest timestamp wins across both arms: both hosts'
-    // :07 rows lead the merged result.
-    assert!(out[0].entry.timestamp.ends_with(":07Z"));
-    assert!(out[1].entry.timestamp.ends_with(":07Z"));
-}
-
-#[test]
-fn mixed_case_hostname_does_not_match_canonical_instance_key() {
-    let (_dir, pool) = test_pool("service-instance-case-miss.db");
-    insert_logs_batch(
-        &pool,
-        &[syslog_row("2026-01-01T00:00:00Z", "Tootie", "plex")],
-    )
-    .unwrap();
-    // Pins the case-sensitivity limitation: canonical keys are lowercase
-    // and the log predicates compare with SQLite's default BINARY
-    // collation, so a mixed-case syslog hostname ("Tootie") never matches
-    // the canonical instance key ("tootie/plex"). Documented in
-    // docs/contracts/investigation-graph.md; hostname case normalization
-    // at ingest is tracked separately.
-    let rows = search_logs_for_service_instances(
-        &pool,
-        &["tootie/plex".to_string()],
-        None,
-        None,
-        None,
-        50,
-    )
-    .unwrap();
-    assert!(rows.is_empty(), "mixed-case hostname must miss (pinned)");
-}
-
-// -- syslog-mcp-csukc: resolve_topic_entities index-backed exact/prefix tier --
-
-#[test]
-fn glob_prefix_pattern_escapes_glob_metacharacters() {
-    // '*', '?' and '[' are GLOB metacharacters; they must be wrapped in a
-    // single-character bracket class so they match literally, not as
-    // wildcards, before the trailing '*' prefix wildcard is appended.
-    assert_eq!(super::glob_prefix_pattern("plex"), "plex*");
-    assert_eq!(super::glob_prefix_pattern("a*b"), "a[*]b*");
-    assert_eq!(super::glob_prefix_pattern("a?b"), "a[?]b*");
-    assert_eq!(super::glob_prefix_pattern("a[b"), "a[[]b*");
-    assert_eq!(super::glob_prefix_pattern("a*b?c[d"), "a[*]b[?]c[[]d*");
-}
-
-#[test]
-fn resolve_topic_entities_exact_and_prefix_use_indexed_tier() {
-    let (_dir, pool) = test_pool("resolve-topic-exact-prefix.db");
-    let conn = pool.get().unwrap();
-    insert_entity(&conn, graph::ENTITY_TYPE_HOST, "tootie");
-    insert_entity(&conn, graph::ENTITY_TYPE_APP, "plex");
-    insert_entity(&conn, graph::ENTITY_TYPE_APP, "plexmediaserver");
-
-    let resolved = super::resolve_topic_entities(&conn, &["plex".to_string()]).unwrap();
-    let exact = resolved
-        .iter()
-        .find(|e| e.canonical_key == "plex")
-        .expect("exact canonical_key match must resolve");
-    assert_eq!(exact.match_kind, "exact");
-    assert_eq!(exact.resolver_status, ResolverStatus::Resolved);
-
-    let prefix = resolved
-        .iter()
-        .find(|e| e.canonical_key == "plexmediaserver")
-        .expect("prefix canonical_key match must resolve");
-    assert_eq!(prefix.match_kind, "prefix");
-    assert_eq!(prefix.resolver_status, ResolverStatus::Ambiguous);
-}
-
-#[test]
-fn resolve_topic_entities_falls_back_to_label_substring_scan() {
-    let (_dir, pool) = test_pool("resolve-topic-label-fallback.db");
-    let conn = pool.get().unwrap();
-    // canonical_key has no relationship to the search term at all; only the
-    // human-readable display_label contains it as a mid-string substring.
-    // This must still be found by the (unavoidable) label scan tier.
-    conn.execute(
-        "INSERT INTO graph_entities
-            (entity_type, canonical_key, display_label, trust_level)
-         VALUES (?1, ?2, ?3, 'verified')",
-        rusqlite::params![
-            graph::ENTITY_TYPE_APP,
-            "backup-job-42",
-            "nightly backup-job-42 (plex library)"
-        ],
-    )
-    .unwrap();
-
-    let resolved = super::resolve_topic_entities(&conn, &["plex".to_string()]).unwrap();
-    let label = resolved
-        .iter()
-        .find(|e| e.canonical_key == "backup-job-42")
-        .expect("label substring match must still resolve via fallback scan");
-    assert_eq!(label.match_kind, "label");
-    assert_eq!(label.resolver_status, ResolverStatus::Ambiguous);
-}
-
-#[test]
-fn resolve_topic_entities_skips_label_scan_when_indexed_tier_fills_cap() {
-    let (_dir, pool) = test_pool("resolve-topic-cap-skips-label.db");
-    let conn = pool.get().unwrap();
-    // 25 canonical_key prefix matches (== PER_TERM_CAP) plus one entity that
-    // would only ever be found via the label substring scan. Once the
-    // indexed tier alone fills the per-term cap, the label tier must not
-    // contribute any additional candidates for this term.
-    for i in 0..25 {
-        conn.execute(
-            "INSERT INTO graph_entities
-                (entity_type, canonical_key, display_label, trust_level)
-             VALUES (?1, ?2, ?2, 'verified')",
-            rusqlite::params![graph::ENTITY_TYPE_APP, format!("svc-{i:02}")],
-        )
-        .unwrap();
-    }
-    conn.execute(
-        "INSERT INTO graph_entities
-            (entity_type, canonical_key, display_label, trust_level)
-         VALUES (?1, ?2, ?3, 'verified')",
-        rusqlite::params![
-            graph::ENTITY_TYPE_APP,
-            "unrelated-key",
-            "totally unrelated but mentions svc- in passing"
-        ],
-    )
-    .unwrap();
-
-    let resolved = super::resolve_topic_entities(&conn, &["svc-".to_string()]).unwrap();
-    assert_eq!(resolved.len(), 25, "per-term cap must still be honored");
-    assert!(
-        !resolved.iter().any(|e| e.canonical_key == "unrelated-key"),
-        "label-only candidate must not appear once the indexed tier fills the cap: {:?}",
-        resolved
     );
 }
